Update main_interface.py
Line 73-116 are the changes, if this isn't exactly what you wanted or there are compatibility issues tell me asap. 
For the time being I'm setting this as a new branch so we can revert to an older iteration if any major issues occur. Delete the created branch and add this update to the master branch if it functions as desired.